+++ src/lib.rs
@@ -1,4 +1,10 @@
 //! MIO bindings for POSIX AIO
+//!
+//! # Feature Flags
+//!
+//! * `tokio` - Add extra methods needed for consumers to implement Tokio's
+//!             `AioSource` trait.
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 mod aio;